Make parent trace names consistent between runs
With tracing in newrelic, groups are done via the name on the parent trace. With each run having a unique name (due to the run ID being unique each time), it makes it hard to use the traces and compare runs over time. If the previous functionality is desired by consumers of this action, it is easy enough to restore by updating the GHA_RUN_NAME environment variable, which is only used in the context of the name of the parent trace.